#34 Migrate launcher CRA to vite
