+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+  "standard.autoFixOnSave": true,
+  "standard.engine": "standardx"
+}
